Extract `statsCompiler`
